feat(motor): 重构电机控制模块并添加PWM电机支持
- 统一使用浮点数处理电机位置和速度数据
- 添加PWM电机类支持占空比控制
- 完善电机数据处理和角度转换逻辑
- 优化CAN通信帧解析和发送代码
- 增加电机位置转换系数配置功能
- 补充数学转换工具函数（角度、弧度、速度等）
- 更新电机基类接口定义和注释文档
- 修复累计位置计算中的数据类型问题
- 移除不必要的头文件包含和变量声明

+++ src/chassis.cpp
@@ -64,7 +64,7 @@ void OmniChassis::NormalizeSpeed(Vector4f& rpm) const {
 void OmniChassis::SetSpeed() {
     Vector4f wheel_rpm=ComputeInverseKinematics(target_vx_, target_vy_, target_omega_);
     for (int i = 0; i < 4; i++) {
-        motors[i]->SetSpeed(static_cast<int>(wheel_rpm(i)));
+        motors[i]->SetSpeed(wheel_rpm(i));
     }
     // if (!device_is_ready(chassis_can_dev)) {
     //     printk("chassis can device is not ready\n");
